+++ src/cli/main.ts
@@ -21,6 +21,7 @@ import {
 } from './client.js';
 import { exitWithError, parsePositiveNumber, printJson } from './common.js';
 import { registerDefaultsCommands } from './defaults-commands.js';
+import { registerPackagesCommands } from './packages-commands.js';
 import { registerServiceCommands } from './service-commands.js';
 
 const defaultApiUrl = process.env.JAGC_API_URL ?? 'http://127.0.0.1:31415';
@@ -52,6 +53,7 @@ program
 
 registerServiceCommands(program);
 registerDefaultsCommands(program);
+registerPackagesCommands(program);
 
 program
   .command('message')